CLION cleanup

+++ memory/dolphin_process.cpp
@@ -42,12 +42,12 @@ dolphin_process::dolphin_process(int _pid) : process{_pid} {
 
         u64 firstAddress = 0;
         u64 SecondAddress = 0;
-        int indexDash = line_data[0].find('-');
+        size_t indexDash = line_data[0].find('-');
         std::string first_address_sring("0x" + line_data[0].substr(0, indexDash));
         std::string second_address_string("0x" + line_data[0].substr(indexDash + 1));
 
-        firstAddress = std::stoul(first_address_sring, nullptr, 16);
-        SecondAddress = std::stoul(second_address_string, nullptr, 16);
+        firstAddress = std::stol(first_address_sring, nullptr, 16);
+        SecondAddress = std::stol(second_address_string, nullptr, 16);
 
         if (SecondAddress - firstAddress == 0x4000000 && offset == 0x2040000) {
             mem2_address = firstAddress;
@@ -63,18 +63,18 @@ dolphin_process::dolphin_process(int _pid) : process{_pid} {
     }
 }
 
-int dolphin_process::read_memory_raw(i64 address, void* buffer, size_t size, int length) {
+size_t dolphin_process::read_memory_raw(u64 address, void* buffer, size_t size, int length) {
     if (address > dolphin_process::gc_memory_start) {
         address -= dolphin_process::gc_memory_start;
     }
-    i64 ram_address = emu_ram_address + address;
+    u64 ram_address = emu_ram_address + address;
     return process::read_memory_raw(ram_address, buffer, size, length);
 }
 
-int dolphin_process::write_memory_raw(i64 address, void* buffer, size_t size, int length) {
+size_t dolphin_process::write_memory_raw(u64 address, void* buffer, size_t size, int length) {
     if (address > dolphin_process::gc_memory_start) {
         address -= dolphin_process::gc_memory_start;
     }
-    i64 ram_address = emu_ram_address + address;
+    u64 ram_address = emu_ram_address + address;
     return process::write_memory_raw(ram_address, buffer, size, length);
 }
